refactor: [#44] extract class GitCommandWrapper
